Pulled `{Rotate,Save}Config` into `BaseFileService` as inner classes.
