+++ src/sys-util/export/BaseFileService.js
@@ -4,9 +4,12 @@
 import * as fs from 'node:fs/promises';
 
 import { WallClock } from '@this/clocks';
+import { BaseConfig } from '@this/compote';
+import { Duration } from '@this/data-values';
 import { Paths, Statter } from '@this/fs-util';
-import { RotateConfig, SaveConfig, ServiceConfig } from '@this/sys-config';
+import { ServiceConfig } from '@this/sys-config';
 import { BaseService } from '@this/sys-framework';
+import { MustBe } from '@this/typey';
 
 
 /**
@@ -76,14 +79,14 @@ export class BaseFileService extends BaseService {
     /**
      * Rotation configuration, if any.
      *
-     * @type {?RotateConfig}
+     * @type {?BaseFileService.RotateConfig}
      */
     #rotate;
 
     /**
      * Preservation configuration, if any.
      *
-     * @type {?SaveConfig}
+     * @type {?BaseFileService.SaveConfig}
      */
     #save;
 
@@ -103,8 +106,8 @@ export class BaseFileService extends BaseService {
       }
 
       this.#path   = Paths.checkAbsolutePath(path);
-      this.#rotate = rotate ? new RotateConfig(rotate) : null;
-      this.#save   = save ? new SaveConfig(save) : null;
+      this.#rotate = rotate ? new BaseFileService.RotateConfig(rotate) : null;
+      this.#save   = save   ? new BaseFileService.SaveConfig(save)     : null;
     }
 
     /**
@@ -141,15 +144,17 @@ export class BaseFileService extends BaseService {
       return this.#pathParts;
     }
 
-    /** @returns {?RotateConfig} Rotation configuration, if any. */
+    /**
+     * @returns {?BaseFileService.RotateConfig} Rotation configuration, if any.
+     */
     get rotate() {
       return this.#rotate;
     }
 
     /**
-     * @returns {?SaveConfig} Preservation configuration, if any. If this
-     * instance has a {@link RotateConfig}, that is returned here too (it's a
-     * subclass).
+     * @returns {?BaseFileService.SaveConfig} Preservation configuration, if
+     * any. If this instance has a {@link #RotateConfig}, that is returned here
+     * too (it's a subclass).
      */
     get save() {
       return this.#save ?? this.#rotate;
@@ -170,6 +175,167 @@ export class BaseFileService extends BaseService {
         fileName.match(/^(?<filePrefix>.*?)(?<fileSuffix>[.][^.]*)?$/).groups;
 
       return Object.freeze({ path, directory, fileName, filePrefix, fileSuffix });
+    }
+  };
+
+  /**
+   * Configuration class for `save` bindings.
+   */
+  static SaveConfig = class SaveConfig extends BaseConfig {
+    /**
+     * The maximum number of old-file bytes to allow, if so limited.
+     *
+     * @type {?number}
+     */
+    #maxOldBytes;
+
+    /**
+     * The maximum number of old files to allow, if so limited.
+     *
+     * @type {?number}
+     */
+    #maxOldCount;
+
+    /**
+     * Rotate when reloading the system?
+     *
+     * @type {boolean}
+     */
+    #onReload;
+
+    /**
+     * Rotate when starting the system?
+     *
+     * @type {boolean}
+     */
+    #onStart;
+
+    /**
+     * Rotate when stopping the system?
+     *
+     * @type {boolean}
+     */
+    #onStop;
+
+    /**
+     * Constructs an instance.
+     *
+     * @param {object} rawConfig Configuration object.
+     */
+    constructor(rawConfig) {
+      super(rawConfig);
+
+      const {
+        maxOldBytes = null,
+        maxOldCount = null,
+        onReload    = false,
+        onStart     = false,
+        onStop      = false
+      } = rawConfig;
+
+      this.#maxOldBytes = (maxOldBytes === null)
+        ? null
+        : MustBe.number(maxOldBytes, { finite: true, minInclusive: 1 });
+      this.#maxOldCount = (maxOldCount === null)
+        ? null
+        : MustBe.number(maxOldCount, { finite: true, minInclusive: 1 });
+      this.#onReload = MustBe.boolean(onReload);
+      this.#onStart  = MustBe.boolean(onStart);
+      this.#onStop   = MustBe.boolean(onStop);
+    }
+
+    /**
+     * @returns {?number} The maximum number of old-file bytes to allow, or
+     * `null` if there is no limit.
+     */
+    get maxOldBytes() {
+      return this.#maxOldBytes;
+    }
+
+    /**
+     * @returns {?number} The maximum number of old files to allow, or `null` if
+     * there is no limit.
+     */
+    get maxOldCount() {
+      return this.#maxOldCount;
+    }
+
+    /** @returns {boolean} Rotate when reloading the system? */
+    get onReload() {
+      return this.#onReload;
+    }
+
+    /** @returns {boolean} Rotate when starting the system? */
+    get onStart() {
+      return this.#onStart;
+    }
+
+    /** @returns {boolean} Rotate when stopping the system? */
+    get onStop() {
+      return this.#onStop;
+    }
+  };
+
+  /**
+   * Configuration class for `rotate` bindings.
+   */
+  static RotateConfig = class RotateConfig extends BaseFileService.SaveConfig {
+    /**
+     * The file size at which to rotate, if ever.
+     *
+     * @type {?number}
+     */
+    #atSize;
+
+    /**
+     * How often to check for rotation eligibility, if at all.
+     *
+     * @type {?Duration}
+     */
+    #checkPeriod;
+
+    /**
+     * Constructs an instance.
+     *
+     * @param {object} rawConfig Configuration object.
+     */
+    constructor(rawConfig) {
+      super(rawConfig);
+
+      const {
+        atSize      = null,
+        checkPeriod = null
+      } = rawConfig;
+
+      this.#atSize = (atSize === null)
+        ? null
+        : MustBe.number(atSize, { finite: true, minInclusive: 1 });
+
+      this.#checkPeriod = Duration.parse(checkPeriod ?? '5 min', { minInclusive: 1 });
+      if (!this.#checkPeriod) {
+        throw new Error(`Could not parse \`checkPeriod\`: ${checkPeriod}`);
+      }
+
+      if (this.#atSize === null) {
+        // `checkPeriod` is irrelevant in this case.
+        this.#checkPeriod = null;
+      }
+    }
+
+    /**
+     * @returns {?number} The file size at which to rotate, or `null` not to do
+     * file size checks.
+     */
+    get atSize() {
+      return this.#atSize;
+    }
+
+    /**
+     * @returns {?Duration} How often to check for rotation eligibility, or
+     * `null` not to ever check.
+     */
+    get checkPeriod() {
+      return this.#checkPeriod;
     }
   };
 }
